Relocated ClientValidator to auth.spi
Since it can be used for both authn and authz
