+++ ms-common-auth/src/main/java/net/trajano/ms/auth/spi/ClientValidator.java
@@ -1,4 +1,4 @@
-package net.trajano.ms.authz.spi;
+package net.trajano.ms.auth.spi;
 
 import java.net.URI;
