+++ ms-common-auth/src/main/java/net/trajano/ms/auth/spi/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Sservice provider interfaces that apply to both authentication and
+ * authorization.
+ *
+ * @author Archimedes Trajano
+ */
+package net.trajano.ms.auth.spi;
