+++ ms-common-auth/src/main/java/net/trajano/ms/authz/TokenResource.java
@@ -35,12 +35,12 @@ import org.springframework.stereotype.Component;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiParam;
+import net.trajano.ms.auth.spi.ClientValidator;
 import net.trajano.ms.auth.token.GrantTypes;
 import net.trajano.ms.auth.token.IdTokenResponse;
 import net.trajano.ms.auth.token.OAuthTokenResponse;
 import net.trajano.ms.auth.util.HttpAuthorizationHeaders;
 import net.trajano.ms.authz.internal.TokenCache;
-import net.trajano.ms.authz.spi.ClientValidator;
 import net.trajano.ms.authz.spi.InternalClaimsBuilder;
 import net.trajano.ms.core.CryptoOps;
 import net.trajano.ms.core.ErrorCodes;
